move compressing classes under lucene90
