+++ lucene/core/src/test/org/apache/lucene/store/TestFileSwitchDirectory.java
@@ -27,7 +27,7 @@ import java.util.HashSet;
 import java.util.Set;
 import java.util.function.Function;
 import org.apache.lucene.analysis.MockAnalyzer;
-import org.apache.lucene.codecs.compressing.Lucene90CompressingStoredFieldsWriter;
+import org.apache.lucene.codecs.lucene90.compressing.Lucene90CompressingStoredFieldsWriter;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexNotFoundException;
 import org.apache.lucene.index.IndexReader;
