+++ lucene/test-framework/src/java/org/apache/lucene/codecs/compressing/CompressingCodec.java
@@ -22,6 +22,8 @@ import org.apache.lucene.codecs.FilterCodec;
 import org.apache.lucene.codecs.StoredFieldsFormat;
 import org.apache.lucene.codecs.TermVectorsFormat;
 import org.apache.lucene.codecs.compressing.dummy.DummyCompressingCodec;
+import org.apache.lucene.codecs.lucene90.compressing.Lucene90CompressingStoredFieldsFormat;
+import org.apache.lucene.codecs.lucene90.compressing.Lucene90CompressingTermVectorsFormat;
 import org.apache.lucene.util.TestUtil;
 import org.apache.lucene.util.packed.DirectMonotonicWriter;
