+++ lucene/test-framework/src/java/org/apache/lucene/codecs/compressing/package.html
@@ -21,6 +21,6 @@
    <meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1">
 </head>
 <body>
-Support for testing {@link org.apache.lucene.codecs.compressing.Lucene90CompressingStoredFieldsFormat}.
+Support for testing {@link org.apache.lucene.codecs.lucene90.compressing.Lucene90CompressingStoredFieldsFormat}.
 </body>
 </html>
